Create Eliz package containing all classes and main
