+++ src/main/java/Eliz/Duke.java
@@ -1,3 +1,5 @@
+package Eliz;
+
 public class Duke {
     public static void main(String[] args) {
         String logo = " ____        _        \n"
