+++ src/main/java/Eliz/Eliz.java
@@ -1,3 +1,5 @@
+package Eliz;
+
 import java.util.Arrays;
 import java.util.Scanner;
