+++ src/main/java/Eliz/ElizException.java
@@ -1,2 +1,4 @@
+package Eliz;
+
 public class ElizException extends Exception{
 }
